requirements.txt
# === Core Python Libraries ===
pandas==2.2.2
numpy==1.26.4

# === Machine Learning ===
scikit-learn==1.5.0
joblib==1.4.2

# === Web Framework & UI ===
Flask==3.0.3
Jinja2==3.1.4
pywebview==5.2

# === Visualization (opsional untuk training/evaluasi) ===
matplotlib==3.9.0
seaborn==0.13.2

# === Google Sheets Integration ===
gspread==6.1.2
oauth2client==4.1.3

# === WhatsApp Notification ===
twilio==9.0.5

# === Utilities ===
requests==2.32.3
python-dotenv==1.0.0